MessageList now only prints 100 messages at a time
Boxes can get very large. A box with several thousand messages could take several seconds to render in MessageList. Now, dbHandler.getMessagesFromMailbox supports a limit and offset so the user doesn't have to wait for the program to retrieve thousands of messages from a very large mailbox before it can render any of them. I've also added a button to MessageList that renders the next 100 messages.